pyproject.toml
[tool.poetry]
name = "infochallengebot"
version = "0.1.0"
description = ""
repository = "https://github.com/UMD-InformationChallenge/InfoChallengeBot"
authors = ["Donal Heidenblad <dheidenb@umd.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.19.2"
validate-email-address = "^1"
SQLAlchemy = "^1.4.29"
# py-cord = { git = "https://github.com/Pycord-Development/pycord.git", rev="4e8b396a", extras = ["speed"]}
py-cord = { version="^2.3.2", extras = ["speed"]}
mariadb = "^1.0.9"

#[tool.poetry.group.docs.dependencies]
#mkdocs = "^1.2.3"

#[tool.poetry.group.utils.dependencies]
#pandas = "^1.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
